requirements.txt
tensorflow
opencv-python
matplotlib
numpy
flask
pillow 
# grad_cam_env\Scripts\activate  -- to activate environment